+++ src/components/productComponents/ProductFiltersComponents/index.ts
@@ -7,4 +7,4 @@ export { default as PriceFilter } from "./PriceFilter";
 export { default as ColorFilter } from "./ColorFilter";
 export { default as MaterialFilter } from "./MaterialFilter";
 export { default as FurnitureCategoryFilter } from "./FurnitureCategoryFilter";
-export { default as RoomCategoryFilter } from "./RoomCategoryFilter";
+export { default as UseProductFilters } from "./UseProductFilters";
